requirements.txt

Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
python-dotenv==1.0.0
